+++ pkg/opensearch/time_series_query_test.go
@@ -10,7 +10,36 @@ import (
 	"github.com/grafana/opensearch-datasource/pkg/tsdb"
 	. "github.com/smartystreets/goconvey/convey"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
+
+func Test_raw_data(t *testing.T) {
+	t.Run("With raw data metric query (from frontend tests)", func(t *testing.T) {
+		from := time.Date(2018, 5, 15, 17, 50, 0, 0, time.UTC)
+		to := time.Date(2018, 5, 15, 17, 55, 0, 0, time.UTC)
+		fromMs := from.UnixNano() / int64(time.Millisecond)
+		toMs := to.UnixNano() / int64(time.Millisecond)
+		c := newFakeClient(es.OpenSearch, "2.3.0")
+
+		_, err := executeTsdbQuery(c, `{
+				"timeField": "@timestamp",
+				"bucketAggs": [],
+				"metrics": [{ "id": "1", "type": "raw_data", "settings": {"size": "1337" }	}]
+			}`, from, to, 15*time.Second)
+		require.NoError(t, err)
+
+		sr := c.multisearchRequests[0].Requests[0]
+		rangeFilter := sr.Query.Bool.Filters[0].(*es.RangeFilter)
+		assert.Equal(t, "@timestamp", rangeFilter.Key)
+		assert.Equal(t, toMs, rangeFilter.Lte)
+		assert.Equal(t, fromMs, rangeFilter.Gte)
+		assert.Equal(t, es.DateFormatEpochMS, rangeFilter.Format)
+
+		assert.Equal(t, 1337, sr.Size)
+		assert.Equal(t, map[string]map[string]string{"@timestamp": {"order": "desc", "unmapped_type": "boolean"}}, sr.Sort[0])
+		assert.Equal(t, map[string]map[string]string{"_doc": {"order": "desc"}}, sr.Sort[1])
+	})
+}
 
 func TestExecuteTimeSeriesQuery(t *testing.T) {
 	from := time.Date(2018, 5, 15, 17, 50, 0, 0, time.UTC)
@@ -262,7 +291,7 @@ func TestExecuteTimeSeriesQuery(t *testing.T) {
 			_, err := executeTsdbQuery(c, `{
 				"timeField": "@timestamp",
 				"bucketAggs": [],
-				"metrics": [{ "id": "1", "type": "raw_document", "settings": { "size": 1337 }	}]
+				"metrics": [{ "id": "1", "type": "raw_document", "settings": { "size": "1337" }	}]
 			}`, from, to, 15*time.Second)
 			So(err, ShouldBeNil)
 			sr := c.multisearchRequests[0].Requests[0]
@@ -807,8 +836,8 @@ func (c *fakeClient) GetVersion() *semver.Version {
 	return c.version
 }
 
-func (c *fakeClient) GetTimeField() string {
-	return c.timeField
+func (c *fakeClient) GetConfiguredFields() es.ConfiguredFields {
+	return es.ConfiguredFields{TimeField: c.timeField}
 }
 
 func (c *fakeClient) GetIndex() string {
@@ -1000,5 +1029,56 @@ func TestTimeSeriesQueryParser(t *testing.T) {
 			So(q.RawQuery, ShouldEqual, "source=index")
 			So(q.QueryType, ShouldEqual, "PPL")
 		})
+	})
+}
+
+func Test_executeTimeSeriesQuery_raw_document_default_size_is_500(t *testing.T) {
+	from := time.Date(2018, 5, 15, 17, 50, 0, 0, time.UTC)
+	to := time.Date(2018, 5, 15, 17, 55, 0, 0, time.UTC)
+	c := newFakeClient(es.OpenSearch, "1.0.0")
+	_, err := executeTsdbQuery(c, `{
+				"timeField": "@timestamp",
+				"bucketAggs": [],
+				"metrics": [{ "id": "1", "type": "raw_document", "settings": {}	}]
+			}`, from, to, 15*time.Second)
+	assert.NoError(t, err)
+	sr := c.multisearchRequests[0].Requests[0]
+
+	assert.Equal(t, 500, sr.Size)
+}
+
+func Test_Field_property(t *testing.T) {
+	from := time.Date(2018, 5, 15, 17, 50, 0, 0, time.UTC)
+	to := time.Date(2018, 5, 15, 17, 55, 0, 0, time.UTC)
+	t.Run("Should use timeField from datasource when not specified", func(t *testing.T) {
+		c := newFakeClient(es.Elasticsearch, "2.0.0")
+		_, err := executeTsdbQuery(c, `{
+			"timeField": "@timestamp",
+			"metrics": [{ "type": "count", "id": "1" }],
+			"bucketAggs": [
+				{ "type": "date_histogram", "id": "2", "settings": { "min_doc_count": "1" } }
+			]
+		}`, from, to, 15*time.Second)
+		assert.Nil(t, err)
+
+		sr := c.multisearchRequests[0].Requests[0]
+		dateHistogramAgg := sr.Aggs[0].Aggregation.Aggregation.(*es.DateHistogramAgg)
+		assert.Equal(t, "@timestamp", dateHistogramAgg.Field)
+	})
+
+	t.Run("Should use field from bucket agg when specified", func(t *testing.T) {
+		c := newFakeClient(es.Elasticsearch, "2.0.0")
+		_, err := executeTsdbQuery(c, `{
+				"timeField": "@timestamp",
+				"metrics": [{ "type": "count", "id": "1" }],
+				"bucketAggs": [
+					{ "type": "date_histogram", "id": "2", "field": "some_other_field", "settings": { "min_doc_count": "1" } }
+				]
+			}`, from, to, 15*time.Second)
+
+		assert.Nil(t, err)
+		sr := c.multisearchRequests[0].Requests[0]
+		dateHistogramAgg := sr.Aggs[0].Aggregation.Aggregation.(*es.DateHistogramAgg)
+		assert.Equal(t, "some_other_field", dateHistogramAgg.Field)
 	})
 }