+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## 2.8.0
+* Fix: Take into account raw_data query's Size and Order  by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/210
+* Backend: Default to timeField if no field is specified in date histogram aggregation by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/215
+* Backend: Change query sort to respect sort order by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/211
+* Backend: Add raw_document query support by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/214
+
+## v2.7.1
+
+* Dependency update
+
+## v2.7.0
+* Add raw_data query support to backend by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/203
+
 ## v2.6.2
 
 * Backend: Convert tables to data frames by @fridgepoet in https://github.com/grafana/opensearch-datasource/pull/186